+++ engine/server/application/CentralServer/src/shared/AuctionTransferGameAPI/Request.cpp
@@ -76,7 +76,7 @@ void ReplyRequest::pack(Base::ByteStream &msg)
 }
 
 //////////////////////////////////////////////////////////////////////////////////////
-CommonRequest::CommonRequest( RequestTypes type, unsigned serverTrack, long long transactionID )
+CommonRequest::CommonRequest( RequestTypes type, unsigned serverTrack, long transactionID )
 	: GenericRequest((short)type, serverTrack), m_transactionID(transactionID)
 {
 }
@@ -101,7 +101,7 @@ void GetIDRequest::pack(Base::ByteStream &msg)
 }
 
 //////////////////////////////////////////////////////////////////////////////////////
-SendPrepareCompressedRequest::SendPrepareCompressedRequest(RequestTypes type, unsigned serverTrack, const char *serverID, long long transactionID, unsigned stationID, unsigned characterID, long long assetID, const unsigned char *xmlAsset, unsigned length)
+SendPrepareCompressedRequest::SendPrepareCompressedRequest(RequestTypes type, unsigned serverTrack, const char *serverID, long transactionID, unsigned stationID, unsigned characterID, long assetID, const unsigned char *xmlAsset, unsigned length)
   : GenericRequest((short)type, serverTrack),  
 	m_transactionID(transactionID), 
 	m_stationID(stationID),
@@ -125,7 +125,7 @@ void SendPrepareCompressedRequest::pack(Base::ByteStream &msg)
 }
 
 //////////////////////////////////////////////////////////////////////////////////////
-SendPrepareRequest::SendPrepareRequest( RequestTypes type, unsigned serverTrack, const char *serverID, long long transactionID, unsigned stationID, unsigned characterID, long long assetID, const char *xml )
+SendPrepareRequest::SendPrepareRequest( RequestTypes type, unsigned serverTrack, const char *serverID, long transactionID, unsigned stationID, unsigned characterID, long assetID, const char *xml )
 	: GenericRequest((short)type, serverTrack),  
 	  m_transactionID(transactionID), 
 	  m_stationID(stationID),
