+++ engine/server/application/CentralServer/src/shared/AuctionTransferGameAPI/Request.h
@@ -43,12 +43,12 @@ namespace AuctionTransfer
 	class CommonRequest: public GenericRequest
 	{
 	public:
-		CommonRequest(RequestTypes type, unsigned serverTrack, long long transactionID);
+		CommonRequest(RequestTypes type, unsigned serverTrack, long transactionID);
 		virtual ~CommonRequest()	{};
 
 		void pack(Base::ByteStream &msg);
 	private:
-		long long m_transactionID;
+		long m_transactionID;
 	};
 
 	//////////////////////////////////////////////////////////////////////////////////////
@@ -66,15 +66,15 @@ namespace AuctionTransfer
 	class SendPrepareRequest: public GenericRequest
 	{
 	public:
-		SendPrepareRequest(RequestTypes type, unsigned serverTrack, const char *serverID, long long transactionID, unsigned stationID, unsigned characterID, long long assetID, const char *xmlAsset);
+		SendPrepareRequest(RequestTypes type, unsigned serverTrack, const char *serverID, long transactionID, unsigned stationID, unsigned characterID, long assetID, const char *xmlAsset);
 		virtual ~SendPrepareRequest()	{};
 
 		void pack(Base::ByteStream &msg);
 	private:
-		long long m_transactionID;
+		long m_transactionID;
 		unsigned m_stationID;
 		unsigned m_characterID;
-		long long m_assetID;
+		long m_assetID;
 		std::string m_xml;
 		std::string m_serverID;
 	};
@@ -83,15 +83,15 @@ namespace AuctionTransfer
 	class SendPrepareCompressedRequest: public GenericRequest
 	{
 	public:
-		SendPrepareCompressedRequest(RequestTypes type, unsigned serverTrack, const char *serverID, long long transactionID, unsigned stationID, unsigned characterID, long long assetID, const unsigned char *xmlAsset, unsigned length);
+		SendPrepareCompressedRequest(RequestTypes type, unsigned serverTrack, const char *serverID, long transactionID, unsigned stationID, unsigned characterID, long assetID, const unsigned char *xmlAsset, unsigned length);
 		virtual ~SendPrepareCompressedRequest()	{};
 
 		void pack(Base::ByteStream &msg);
 	private:
-		long long m_transactionID;
+		long m_transactionID;
 		unsigned m_stationID;
 		unsigned m_characterID;
-		long long m_assetID;
+		long m_assetID;
 		//std::string m_xml;
 		std::string m_serverID;
 		Blob m_data;
@@ -133,7 +133,7 @@ namespace AuctionTransfer
 	private:
 		std::string m_gameCode;
 		std::string m_serverCode;
-		long long m_assetID;
+		long m_assetID;
 		unsigned m_userID;
 		std::string m_event;
 		std::string m_message;
